Main Program code.
The simulator logic is functioning, but linking it to the qt5 gui has proven to be an issue. This is the first time I've looked at the QT5 library and threading didn't work as I had expected. It seems the main logic gets out of sync with the gui and results in a crash. I will continue to look at this further.